修改分页查询，增加全局异常

+++ src/main/java/com/sge/service/PlatformService.java
@@ -7,10 +7,10 @@ import com.sge.repository.BranchRepository;
 import com.sge.repository.PlatformRepository;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Example;
+import org.springframework.data.domain.*;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
+import java.sql.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -24,13 +24,30 @@ public class PlatformService {
 
     @Autowired
     private PlatformRepository platformRepository;
+    @Autowired
+    private BranchRepository branchRepository;
 
     /**
      * 查询
      * @return
+     * @param sysCode
+     * @param pageSize
+     * @param pageNum
      */
-    public List<Platform> findAll(){
-        return platformRepository.findAll();
+    public Page<Platform> findByPage(String sysCode, int pageSize, int pageNum){
+        Sort sort =  Sort.by(Sort.Direction.DESC,"createTime");
+        Pageable pageable = PageRequest.of(pageNum-1,pageSize,sort);
+        Page<Platform> page = null;
+        if (StringUtils.isBlank(sysCode)) {
+            page = platformRepository.findAll(pageable);
+        } else {
+            Platform query = new Platform();
+            query.setSystemCode(sysCode);
+            Example<Platform> example = Example.of(query);
+            page = platformRepository.findAll(example,pageable);
+        }
+
+        return page;
     }
 
     /**
@@ -39,6 +56,22 @@ public class PlatformService {
      */
     public void save(Platform platform) {
         validateParam(platform);
+        Platform codeValid = new Platform();
+        codeValid.setSystemCode(platform.getSystemName());
+        Example<Platform> exampleCode = Example.of(codeValid);
+        long countCode = platformRepository.count(exampleCode);
+        if (countCode > 0) {
+            throw new BusinessException("系统编码不能重复");
+        }
+        Platform nameValid = new Platform();
+        nameValid.setSystemName(platform.getSystemName());
+        Example<Platform> exampleName = Example.of(nameValid);
+        long countName = platformRepository.count(exampleName);
+        if (countName > 0) {
+            throw new BusinessException("系统名称不能重复");
+        }
+        platform.setCreateTime(new Date(System.currentTimeMillis()));
+        platform.setUpdateTime(new Date(System.currentTimeMillis()));
         platformRepository.save(platform);
     }
 
@@ -50,8 +83,8 @@ public class PlatformService {
         if (StringUtils.isBlank(platform.getSystemName())) {
             throw new BusinessException("请输入系统名称");
         }
-        if (StringUtils.isBlank(platform.getDesc())) {
-            throw new BusinessException("请输入系统描述");
+        if (StringUtils.isBlank(platform.getSystemCode())) {
+            throw new BusinessException("请输入系统编码");
         }
     }
 
@@ -63,6 +96,13 @@ public class PlatformService {
         boolean flag = platformRepository.existsById(id);
         if (!flag) {
             throw new BusinessException("该记录不存在");
+        }
+        Branch branch = new Branch();
+        branch.setPlatformId(id);
+        Example<Branch> example= Example.of(branch);
+        long count = branchRepository.count(example);
+        if (count > 0) {
+            throw new BusinessException("该系统存在分支不能删除");
         }
         platformRepository.deleteById(id);
     }
@@ -76,7 +116,32 @@ public class PlatformService {
         if (!flag) {
             throw new BusinessException("该记录不存在");
         }
+        validateParam(platform);
+        int id = platform.getId();
+        String name = platform.getSystemName();
+        String code = platform.getSystemCode();
+        int nameCount = platformRepository.countByNameOrCode(id,null,code);
+        if (nameCount > 0) {
+            throw new BusinessException("系统编码不能重复");
+        }
+        int codeCount = platformRepository.countByNameOrCode(id,name,null);
+        if (codeCount > 0) {
+            throw new BusinessException("系统名称不能重复");
+        }
         platformRepository.save(platform);
     }
 
+    public List<Platform> findAll() {
+        return platformRepository.findAll();
+    }
+
+    /**
+     * 根据主键查询
+     * @param id
+     * @return
+     */
+    public Platform getById(Integer id) {
+        Optional<Platform> optional = platformRepository.findById(id);
+        return optional.get();
+    }
 }
